fix(DEV-4216): type exchange_ratio as number in COIN reward schemas
Regenerated SDK from updated OpenAPI after exchange_ratio integer/string
was changed to number to support fractional values such as 0.01.

+++ src/Model/RedemptionRewardResultParametersCoin.php
@@ -59,7 +59,7 @@ class RedemptionRewardResultParametersCoin implements ModelInterface, ArrayAcces
       * @var string[]
       */
     protected static $openAPITypes = [
-        'exchangeRatio' => 'int',
+        'exchangeRatio' => 'float',
         'pointsRatio' => 'int'
     ];
 
@@ -301,7 +301,7 @@ class RedemptionRewardResultParametersCoin implements ModelInterface, ArrayAcces
     /**
      * Gets exchangeRatio
      *
-     * @return int|null
+     * @return float|null
      */
     public function getExchangeRatio()
     {
@@ -311,7 +311,7 @@ class RedemptionRewardResultParametersCoin implements ModelInterface, ArrayAcces
     /**
      * Sets exchangeRatio
      *
-     * @param int|null $exchangeRatio The cash equivalent of the points defined in the `points_ratio` property.
+     * @param float|null $exchangeRatio The cash equivalent of the points defined in the `points_ratio` property.
      *
      * @return self
      */
